requirements.txt
aiohttp==3.7.4.post0
<<<<<<< HEAD
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cytoolz==0.11.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.1
django-cors-headers==3.7.0
django-dotenv==1.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
gunicorn==20.1.0
hexbytes==0.2.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.0
incremental==21.3.0
ipfshttpclient==0.7.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
parsimonious==0.8.1
Pillow==8.2.0
protobuf==3.17.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyrsistent==0.17.3
pytz==2021.1
requests==2.25.1
rlp==2.0.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toolz==0.11.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
web3==5.20.0
websockets==8.1
whitenoise==5.3.0
yarl==1.6.3
zipp==3.4.1
zope.interface==5.4.0
=======
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
async-timeout==3.0.1
attrdict==2.0.1
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bitarray==1.2.2
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==35.0.0
cycler==0.10.0
cytoolz==0.11.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
ethereum-keys==0.1.0a9
ethereum-utils==0.6.2
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.2
importlib-metadata==4.8.1
ipfshttpclient==0.8.0a2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lru-dict==1.1.7
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
messagebird==2.0.0
mistune==0.8.4
multiaddr==0.0.9
multidict==5.1.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netaddr==0.8.0
notebook==6.4.4
null==0.6.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
py-solc==3.2.0
py-solc-x==1.1.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
requests==2.26.0
rlp==1.2.0
scikit-learn==0.24.2
scipy==1.7.1
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.16.0
solc==0.0.0a0
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
toolz==0.11.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
varint==1.0.2
wcwidth==0.2.5
web3==5.23.1
webencodings==0.5.1
websockets==9.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
yarl==1.6.3
zipp==3.5.0
>>>>>>> origin/main
